+++ src/main/java/nl/pallett/jsoneditor/actions/ActionManager.java
@@ -8,6 +8,9 @@ import nl.pallett.jsoneditor.actions.file.SaveAsAction;
 import nl.pallett.jsoneditor.actions.find.ShowFindDialogAction;
 import nl.pallett.jsoneditor.actions.find.ShowGotoLineDialogAction;
 import nl.pallett.jsoneditor.actions.find.ShowReplaceDialogAction;
+import nl.pallett.jsoneditor.actions.view.DecreaseFontSizeAction;
+import nl.pallett.jsoneditor.actions.view.IncreaseFontSizeAction;
+import nl.pallett.jsoneditor.actions.view.ResetFontSizeAction;
 import nl.pallett.jsoneditor.controller.EditorManager;
 import nl.pallett.jsoneditor.view.editor.EditorPanelView;
 import org.jspecify.annotations.Nullable;
@@ -33,6 +36,9 @@ public class ActionManager {
         actions.put(Action.FIND, new ShowFindDialogAction(editorManager));
         actions.put(Action.REPLACE, new ShowReplaceDialogAction(editorManager));
         actions.put(Action.GOTO_LINE, new ShowGotoLineDialogAction(editorManager));
+        actions.put(Action.INCREASE_FONT_SIZE, new IncreaseFontSizeAction(editorManager));
+        actions.put(Action.DECREASE_FONT_SIZE, new DecreaseFontSizeAction(editorManager));
+        actions.put(Action.RESET_FONT_SIZE, new ResetFontSizeAction(editorManager));
     }
 
     public void updateState(@Nullable EditorPanelView editorPanel) {
@@ -59,7 +65,10 @@ public class ActionManager {
         REDO,
         FIND,
         REPLACE,
-        GOTO_LINE
+        GOTO_LINE,
+        INCREASE_FONT_SIZE,
+        DECREASE_FONT_SIZE,
+        RESET_FONT_SIZE
     }
 
 }